third_party/WebKit/LayoutTests/geolocation-api/error-service-connection-error-expected.txt

Tests Geolocation when the geolocation service connection fails.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User denied Geolocation"
PASS successfullyParsed is true

TEST COMPLETE
